apply the same folder flat class structure to preprocessor
